6-size.c1

+++ hello_world/6-size.c
@@ -13,7 +13,7 @@ float floatType;
 double doubleType;
 char charType;
 printf("Size of Char Data Types in C = %2d bytes \n", sizeof(char));
-printf("Size of Int Data Types in C = %2d bytes \n", sizeof(short int));
+printf("Size of Int Data Types in C = %2d bytes \n", sizeof(int));
 printf("Size of Long Int Data Types in C = %2d bytes \n", sizeof(long int));
 printf("Size of Long Double Data Types in C = %2d bytes \n", sizeof(double));
 printf("Size of Float Data Types in C = %2d bytes \n", sizeof(float));
